express-validator implementation with unit testing and user model chenged
