+++ Backend/src/controllers/index.ts
@@ -1,2 +1,3 @@
+export * from './auth.controller';
 export * from './database.controller';
 export * from './users.controller';
