+++ Backend/src/routes/auth.routes.ts
@@ -1,14 +1,43 @@
 import { Router } from "express";
-import { createTableUsers, login, registerUser } from "../controllers";
-import { body } from "express-validator";
+import { login, registerUser } from "../controllers";
+import { check } from "express-validator";
 const router: Router = Router();
 
+const passRegex: RegExp = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])[0-9a-zA-Z]{8,}$/;
 
-router.post("/register", registerUser);
+/**
+ * ^      -> start of the input
+ * (?=)   -> first condition
+ * .      -> any character except newline (\n)
+ * *\d    -> zero or more times any character between [0-9]
+ * (?=)   -> second condition
+ * .      -> any character except newline (\n)
+ * *[a-z] -> zero or more times any character between [a-z]
+ * (?=)   -> third condition
+ * .      -> any character except newline (\n)
+ * *[A-Z] -> zero or more times any character between [A-Z]
+ * [0-9a-zA-Z] -> following characters must be any of these characters
+ * {8,}$  -> 8 or more characters at the end
+ * 
+ * 
+ * example: "92johnDOE4ever"
+ */
+
+
+router.post("/register", 
+  check('email').isEmail().withMessage("Email must have a valid format"),
+  check("password").isLength({ min: 6 }).withMessage("Password must be longer than 6 characters"),
+  check('password').matches(passRegex, "i").withMessage("Please enter a valid password"),
+  check('role').isString().withMessage("Role must have a valid value"),
+  check('role').custom((value) => value==='admin' || value==='employee').withMessage("Role must be a valid one")
+,registerUser);
+
+
 router.post(
   "/login",
-  body("email").isEmail().normalizeEmail(),
-  body("password").isLength({ min: 6 }),
+  check('email').isEmail().withMessage("Email must have a valid format"),
+  check("password").isLength({ min: 6 }).withMessage("Password must be longer than 6 characters"),
+  check('password').matches(passRegex, "i").withMessage("Please enter a valid password"),
   login
 );
