+++ Backend/src/types/types.d.ts
@@ -2,7 +2,8 @@ export interface UserBody {
   name: string,
   second_name: string,
   email: string,
-  password: string
+  password: string,
+  role: string
 }
 
 export interface UserLogin {
